Comment out any properties that have a 'nil' value
This is mainly needed for a 5.6+ upgrade that I intent to do. I noticed
that starting a 5.6.6 SonarQube with the old configuration would use the
embedded database even though I had configured a proper JDBC url.
Commenting out some of the embedded database configuration actually made
sure it used the configured (MySQL) jdbc url.

All configuration properties that are now set to 'nil' will be commented
out (but still visible in the configuration file)